Feature/user permissions

+++ api/src/Controller/ParkingController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 
 use App\Entity\Parkings;
 use App\Entity\Users;
+use App\EventListener\JWTDecodedListener;
 use App\Util\EncodeJSON;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -89,10 +90,12 @@ class ParkingController extends AbstractController
         $data = $request->getContent();
         $entityManager = $this->doctrine->getManager();
         $parking = $entityManager->getRepository(Parkings::class)->find($id);
-
         if (!$parking) {
             return $this->json("No parking found for id: $id", 404);
         }
+
+        $this->denyAccessUnlessGranted("edit", $parking);
+
         $parkingDecode = json_decode($data);
         $updateParking = EncodeJSON::DecodeParking($parkingDecode, $parking, true);
 
@@ -118,6 +121,7 @@ class ParkingController extends AbstractController
         if (!$parking) {
             return $this->json("No parking found for id: $id", 404);
         }
+        $this->denyAccessUnlessGranted("edit", $parking);
 
         $entityManager->remove($parking);
         $entityManager->flush();
