+++ api/src/Controller/UserController.php
@@ -8,7 +8,6 @@ use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
@@ -31,6 +30,8 @@ class UserController extends AbstractController
         if (!$user) {
             return $this->json("No user found for id: $id", 404);
         }
+        // return sensitive information about the user, only the user can see their information.
+        $this->denyAccessUnlessGranted("view", $user);
         // return the full information of the user
         $data = EncodeJSON::EncodeUser($user);
 
@@ -38,21 +39,52 @@ class UserController extends AbstractController
     }
 
     /**
+     * @Route("/api/users", name="users", methods={"GET"})
+     */
+    public function allUsers(): Response
+    {
+        $this->denyAccessUnlessGranted("ROLE_SUPER_ADMIN");
+        $entityManager = $this->doctrine->getManager();
+        $users = $entityManager->getRepository(Users::class)->findAll();
+        if (!$users) {
+            return $this->json("No user found", 404);
+        }
+        // return sensitive information about the user, only the user can see their information.
+//        $this->denyAccessUnlessGranted("view", $user);
+        // return the full information of the user
+        $data = [];
+        foreach ($users as $user) {
+            $data[] = EncodeJSON::EncodeUser($user);
+        }
+
+        return $this->json($data);
+    }
+
+    /**
      * @Route("/api/user", name="userid", methods={"GET"})
+     * get method can't have body, so the mail is passed by a query
      */
     public function getIdOfUser(Request $request): Response
     {
-        $mail = $request->query->get('mail');
-        $entityManager = $this->doctrine->getManager();
-        $user = $entityManager->getRepository(Users::class)->findOneByMail($mail);
+//        $mail = $request->query->get('mail');
+//        $mail =
+//        $entityManager = $this->doctrine->getManager();
+//        $user = $entityManager->getRepository(Users::class)->findOneByMail($mail);
+//        if (!$user) {
+//            return $this->json("No user found for mail: $mail", 404);
+//        }
+
+        $user = $this->getUser();
         if (!$user) {
-            return $this->json("No user found for mail: $mail", 404);
+            return $this->json("No user found", 404);
         }
+
         // return the full information of the user
         $data = EncodeJSON::EncodeUser($user, true, true);
 
         return $this->json($data);
     }
+
 
     /**
      * @Route("/api/user/{id}", name="updateUser", methods={"PUT"}, requirements={"id": "\d+"})
@@ -60,12 +92,15 @@ class UserController extends AbstractController
     public function updateUser(int $id, Request $request): Response
     {
         $data = $request->getContent();
+
         // update user info, no password change
         $entityManager = $this->doctrine->getManager();
         $user = $entityManager->getRepository(Users::class)->find($id);
         if (!$user) {
             return $this->json("No user found for id: $id", 404);
         }
+        $this->denyAccessUnlessGranted("edit", $user);
+
         $userDecode = json_decode($data);
         $updateUser = EncodeJSON::DecodeUser($userDecode, false, true, $user);
         $errors = $this->validator->validate($updateUser);
@@ -92,6 +127,7 @@ class UserController extends AbstractController
         if (!$user) {
             return $this->json("No user found for id: $id", 404);
         }
+        $this->denyAccessUnlessGranted("edit", $user);
 
         $entityManager->remove($user);
         $entityManager->flush();
